requirements-adapter.txt
adapter-transformers==3.1.0
torch==1.8.0
higher==0.2.1
torchtext==0.6.0
scikit-learn==0.23.2
pandas==1.1.0
numpy==1.23.5